+++ packages/strapi-design-system/src/Table/Table.stories.mdx
@@ -5,7 +5,8 @@ import { withDesign } from 'storybook-addon-designs';
 import { Table } from './Table';
 import { Th, Td } from './Cell';
 import { Tr } from './Tr';
-import { Thead, Tbody } from './Content';
+import { Thead } from './Thead';
+import { Tbody } from './Tbody';
 import { Box } from '../Box';
 
 <Meta
