+++ packages/strapi-design-system/src/Table/Tbody.js
@@ -2,19 +2,6 @@ import React, { cloneElement, Children } from 'react';
 import PropTypes from 'prop-types';
 import { useTable } from './TableContext';
 
-export const Thead = ({ children, ...props }) => {
-  const { rowIndex, colIndex } = useTable();
-
-  /**
-   * aria-rowindex is 1-based: we have to start from 1
-   */
-  const childrenClone = Children.toArray(children).map((child, index) =>
-    cloneElement(child, { focusedColIndex: rowIndex === index ? colIndex : undefined, 'aria-rowindex': 1 }),
-  );
-
-  return <thead {...props}>{childrenClone}</thead>;
-};
-
 export const Tbody = ({ children, ...props }) => {
   const { rowIndex, colIndex } = useTable();
 
@@ -27,10 +14,6 @@ export const Tbody = ({ children, ...props }) => {
   );
 
   return <tbody {...props}>{childrenClone}</tbody>;
-};
-
-Thead.propTypes = {
-  children: PropTypes.node.isRequired,
 };
 
 Tbody.propTypes = {
