+++ packages/strapi-design-system/src/Table/Thead.js
@@ -1,0 +1,20 @@
+import React, { cloneElement, Children } from 'react';
+import PropTypes from 'prop-types';
+import { useTable } from './TableContext';
+
+export const Thead = ({ children, ...props }) => {
+  const { rowIndex, colIndex } = useTable();
+
+  /**
+   * aria-rowindex is 1-based: we have to start from 1
+   */
+  const childrenClone = Children.toArray(children).map((child, index) =>
+    cloneElement(child, { focusedColIndex: rowIndex === index ? colIndex : undefined, 'aria-rowindex': 1 }),
+  );
+
+  return <thead {...props}>{childrenClone}</thead>;
+};
+
+Thead.propTypes = {
+  children: PropTypes.node.isRequired,
+};
